+++ jest.setup.js
@@ -3,3 +3,6 @@ import mockRNCNetInfo from '@react-native-community/netinfo/jest/netinfo-mock';
 
 jest.mock('@react-native-async-storage/async-storage', () => mockAsyncStorage);
 jest.mock('@react-native-community/netinfo', () => mockRNCNetInfo);
+jest.mock('react-native-fs', () => ({
+  readFile: jest.fn(),
+}));
