+++ package.json
@@ -60,6 +60,7 @@
     "react-native-background-fetch": "3.1.0",
     "react-native-base64": "^0.2.1",
     "react-native-config": "^1.4.2",
+    "react-native-fs": "^2.16.6",
     "react-native-gesture-handler": "1.8.0",
     "react-native-linear-gradient": "^2.5.6",
     "react-native-localize": "^1.4.0",
