+++ src/App.tsx
@@ -24,7 +24,7 @@ import {NotificationPermissionStatusProvider} from 'shared/NotificationPermissio
 
 import regionContentDefault from './locale/translations/region.json';
 import {RegionContent, RegionContentResponse} from './shared/Region';
-import {OutbreakProvider} from './shared/OutbreakProvider';
+import {OutbreakProvider} from './services/OutbreakService';
 
 // this allows us to use new Date().toLocaleString() for date formatting on android
 // https://github.com/facebook/react-native/issues/19410#issuecomment-482804142
@@ -82,7 +82,7 @@ const App = () => {
     <I18nProvider>
       <RegionalProvider activeRegions={[]} translate={id => id} regionContent={regionContent.payload}>
         <ExposureNotificationServiceProvider backendInterface={backendService}>
-          <OutbreakProvider>
+          <OutbreakProvider backendService={backendService}>
             <DevPersistedNavigationContainer>
               <AccessibilityServiceProvider>
                 <NotificationPermissionStatusProvider>
