+++ src/screens/home/HomeScreen.tsx
@@ -19,7 +19,7 @@ import {getRegionCase} from 'shared/RegionLogic';
 import {ForceScreen} from 'shared/ForceScreen';
 import {useRegionalI18n} from 'locale';
 import {isExposedToOutbreak} from 'shared/qr';
-import {useOutbreakService} from 'shared/OutbreakProvider';
+import {useOutbreakService} from 'services/OutbreakService';
 import {useNotificationPermissionStatus} from 'shared/NotificationPermissionStatus';
 
 import {useDeepLinks} from '../qr/utils';
