+++ src/screens/qr/utils.ts
@@ -6,7 +6,7 @@ import {log} from 'shared/logging/config';
 import {useNavigation} from '@react-navigation/native';
 import {CheckInData} from 'shared/qr';
 import {getCurrentDate} from 'shared/date-fns';
-import {useOutbreakService} from 'shared/OutbreakProvider';
+import {useOutbreakService} from 'services/OutbreakService';
 import {QR_HOST, QR_CODE_PUBLIC_KEY} from 'env';
 import base64 from 'react-native-base64';
 import nacl from 'tweetnacl';
