+++ src/screens/qr/views/QRCodeScanner.tsx
@@ -6,7 +6,7 @@ import {Box, Text, Toolbar2} from 'components';
 import {useI18n} from 'locale';
 import {useNavigation} from '@react-navigation/native';
 import {log} from 'shared/logging/config';
-import {useOutbreakService} from 'shared/OutbreakProvider';
+import {useOutbreakService} from 'services/OutbreakService';
 
 import {handleOpenURL} from '../utils';
